Move createDocument to TestCase
